+++ osu-Replay-Analyzer/osu!/Drawable/OsuStdRenderer.cpp
@@ -4,7 +4,7 @@
 
 #include "../osu_standard.h"
 
-OsuStdRenderer::OsuStdRenderer(Play* _play, int* _viewTime, GuiObj* _parent) : GuiObj(0, 0, _parent->getDim().Width, _parent->getDim().Height, _parent)
+OsuStdRenderer::OsuStdRenderer(Play* _play, long* _viewTime, GuiObj* _parent) : GuiObj(0, 0, _parent->getDim().Width, _parent->getDim().Height, _parent)
 {
 	play = _play;
 	viewTime = _viewTime;
